refactor: remove GateConfig, legacy condition types, and entry/exit gate UI
Delete GateConfig.tsx (legacy gate type selector) and
legacy-condition-types.ts. Remove entry/exit gate UI from
WorkflowNodeCard — these were frontend-only artifacts with no
runtime backing (channel gates replaced edge transitions).

Remove Shell Condition from ChannelEdgeConfigPanel gate mode list.
