+++ packages/web/src/components/space/visual-editor/EdgeRenderer.tsx
@@ -31,8 +31,7 @@
  */
 
 import { useEffect, useRef } from 'preact/hooks';
-import type { WorkflowConditionType } from './legacy-condition-types';
-import type { NodePosition, VisualTransition } from './types';
+import type { NodePosition, VisualTransition, WorkflowConditionType } from './types';
 import type { AnchorSide } from './semanticWorkflowGraph';
 
 // Module-level counter -- increments on each EdgeRenderer mount, giving every
