+++ packages/web/src/components/space/visual-editor/serialization.ts
@@ -35,8 +35,7 @@ import type {
 import type { NodeDraft } from '../WorkflowNodeCard';
 import type { RuleDraft } from '../WorkflowRulesEditor';
 import { rulesToDrafts } from '../WorkflowRulesEditor';
-import type { WorkflowCondition } from './legacy-condition-types';
-import type { Point } from './types';
+import type { Point, WorkflowCondition } from './types';
 import { autoLayout } from './layout';
 
 // ============================================================================
